+++ src/styles/scene/macbook/bootScreen.module.scss
@@ -1,0 +1,6 @@
+.container {
+    width: 100%;
+    height: 100vh;
+
+    background-color: var(--color-mac-boot-screen);
+}
