--- src/type/mesh.ts
@@ -1,6 +1,0 @@
-import * as THREE from 'three';
-
-export interface MeshWithMaterial extends THREE.Object3D {
-	material?: THREE.Material;
-	geometry?: THREE.BufferGeometry;
-}
